refactor: move history cli to its own module
